move ui demo to separate model
