+++ settings.gradle
@@ -14,6 +14,7 @@ include 'selfpayments'
 include 'statistics'
 include 'transactions'
 include 'web'
+include 'ui-demo'
 include 'grpc-adapter'
 include 'grpc-client'
 include 'application'
