+++ ui-demo/src/main/java/de/cotto/lndmanagej/ui/demo/DemoApplication.java
@@ -1,6 +1,7 @@
-package de.cotto.lndmanagej;
+package de.cotto.lndmanagej.ui.demo;
 
-import de.cotto.lndmanagej.demo.DemoDataService;
+import de.cotto.lndmanagej.controller.ChannelIdConverter;
+import de.cotto.lndmanagej.controller.PubkeyConverter;
 import de.cotto.lndmanagej.ui.UiDataService;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
@@ -10,6 +11,7 @@ import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfigurat
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 
 @Configuration
 @EnableAutoConfiguration(
@@ -20,6 +22,7 @@ import org.springframework.context.annotation.Configuration;
         }
 )
 @ComponentScan(basePackages = {"de.cotto.lndmanagej.ui"})
+@Import({ChannelIdConverter.class, PubkeyConverter.class})
 public class DemoApplication {
 
     public DemoApplication() {
