+++ ui-demo/src/main/resources/application.properties
