+++ web/src/testFixtures/java/de/cotto/lndmanagej/controller/dto/NodeDetailsDtoFixture.java
@@ -1,4 +1,4 @@
-package de.cotto.lndmanagej.ui.model;
+package de.cotto.lndmanagej.controller.dto;
 
 import de.cotto.lndmanagej.controller.dto.BalanceInformationDto;
 import de.cotto.lndmanagej.controller.dto.FeeReportDto;
@@ -7,9 +7,11 @@ import de.cotto.lndmanagej.controller.dto.NodeDetailsDto;
 import de.cotto.lndmanagej.controller.dto.OnChainCostsDto;
 import de.cotto.lndmanagej.controller.dto.OnlineReportDto;
 import de.cotto.lndmanagej.controller.dto.RebalanceReportDto;
+import de.cotto.lndmanagej.model.BalanceInformationFixtures;
 import de.cotto.lndmanagej.model.ChannelIdFixtures;
 import de.cotto.lndmanagej.model.FeeReportFixtures;
 import de.cotto.lndmanagej.model.FlowReportFixtures;
+import de.cotto.lndmanagej.model.OnChainCostsFixtures;
 import de.cotto.lndmanagej.model.OnlineReportFixtures;
 import de.cotto.lndmanagej.model.PubkeyFixtures;
 import de.cotto.lndmanagej.model.RebalanceReportFixtures;
@@ -18,6 +20,7 @@ import de.cotto.lndmanagej.model.warnings.ChannelWarningsFixtures;
 import java.util.List;
 
 import static de.cotto.lndmanagej.model.BalanceInformationFixtures.BALANCE_INFORMATION;
+import static de.cotto.lndmanagej.model.ChannelIdFixtures.CHANNEL_ID;
 import static de.cotto.lndmanagej.model.FeeReportFixtures.FEE_REPORT;
 import static de.cotto.lndmanagej.model.FlowReportFixtures.FLOW_REPORT;
 import static de.cotto.lndmanagej.model.OnChainCostsFixtures.ON_CHAIN_COSTS;
@@ -29,12 +32,12 @@ import static de.cotto.lndmanagej.model.warnings.ChannelWarningsFixtures.CHANNEL
 public class NodeDetailsDtoFixture {
 
     public static final NodeDetailsDto NODE_DETAILS_DTO = new NodeDetailsDto(
-            PUBKEY,
+            PubkeyFixtures.PUBKEY,
             "Albert",
-            List.of(ChannelIdFixtures.CHANNEL_ID),
-            List.of(ChannelIdFixtures.CHANNEL_ID),
-            List.of(ChannelIdFixtures.CHANNEL_ID),
-            List.of(ChannelIdFixtures.CHANNEL_ID),
+            List.of(CHANNEL_ID),
+            List.of(CHANNEL_ID),
+            List.of(CHANNEL_ID),
+            List.of(CHANNEL_ID),
             OnChainCostsDto.createFromModel(ON_CHAIN_COSTS),
             BalanceInformationDto.createFromModel(BALANCE_INFORMATION),
             OnlineReportDto.createFromModel(ONLINE_REPORT),
