+++ web/src/testFixtures/java/de/cotto/lndmanagej/controller/dto/OpenChannelDtoFixture.java
@@ -1,31 +1,35 @@
-package de.cotto.lndmanagej.ui.model;
+package de.cotto.lndmanagej.controller.dto;
 
 import de.cotto.lndmanagej.controller.dto.BalanceInformationDto;
 import de.cotto.lndmanagej.controller.dto.PoliciesDto;
 import de.cotto.lndmanagej.model.BalanceInformation;
 import de.cotto.lndmanagej.model.ChannelId;
-import de.cotto.lndmanagej.model.ChannelIdFixtures;
 import de.cotto.lndmanagej.model.Coins;
 import de.cotto.lndmanagej.model.Pubkey;
-import de.cotto.lndmanagej.model.PubkeyFixtures;
 import de.cotto.lndmanagej.ui.dto.OpenChannelDto;
 
 import static de.cotto.lndmanagej.model.BalanceInformationFixtures.BALANCE_INFORMATION;
+import static de.cotto.lndmanagej.model.ChannelIdFixtures.CHANNEL_ID;
+import static de.cotto.lndmanagej.model.ChannelIdFixtures.CHANNEL_ID_2;
+import static de.cotto.lndmanagej.model.ChannelIdFixtures.CHANNEL_ID_3;
+import static de.cotto.lndmanagej.model.ChannelIdFixtures.CHANNEL_ID_4;
+import static de.cotto.lndmanagej.model.ChannelIdFixtures.CHANNEL_ID_5;
 import static de.cotto.lndmanagej.model.PolicyFixtures.POLICIES_FOR_LOCAL_CHANNEL;
+import static de.cotto.lndmanagej.model.PubkeyFixtures.PUBKEY;
 
 public class OpenChannelDtoFixture {
 
     public static final long MILLION = 1_000_000;
 
     public static final OpenChannelDto OPEN_CHANNEL_DTO = new OpenChannelDto(
-            ChannelIdFixtures.CHANNEL_ID,
+            CHANNEL_ID,
             "Albert",
-            PubkeyFixtures.PUBKEY,
+            PUBKEY,
             PoliciesDto.createFromModel(POLICIES_FOR_LOCAL_CHANNEL),
             BalanceInformationDto.createFromModel(BALANCE_INFORMATION));
 
     public static final OpenChannelDto ACINQ = new OpenChannelDto(
-            ChannelIdFixtures.CHANNEL_ID_2,
+            CHANNEL_ID_2,
             "ACINQ",
             Pubkey.create("03864ef025fde8fb587d989186ce6a4a186895ee44a926bfc370e2c366597a3f8f"),
             PoliciesDto.createFromModel(POLICIES_FOR_LOCAL_CHANNEL),
@@ -49,7 +53,7 @@ public class OpenChannelDtoFixture {
             )));
 
     public static final OpenChannelDto WOS = new OpenChannelDto(
-            ChannelIdFixtures.CHANNEL_ID_3,
+            CHANNEL_ID_3,
             "WalletOfSatoshi.com",
             Pubkey.create("035e4ff418fc8b5554c5d9eea66396c227bd429a3251c8cbc711002ba215bfc226"),
             PoliciesDto.createFromModel(POLICIES_FOR_LOCAL_CHANNEL),
@@ -61,7 +65,7 @@ public class OpenChannelDtoFixture {
             )));
 
     public static final OpenChannelDto WOS2 = new OpenChannelDto(
-            ChannelIdFixtures.CHANNEL_ID_3,
+            CHANNEL_ID_3,
             "WalletOfSatoshi.com",
             Pubkey.create("035e4ff418fc8b5554c5d9eea66396c227bd429a3251c8cbc711002ba215bfc226"),
             PoliciesDto.createFromModel(POLICIES_FOR_LOCAL_CHANNEL),
@@ -73,7 +77,7 @@ public class OpenChannelDtoFixture {
             )));
 
     public static final OpenChannelDto BCASH = new OpenChannelDto(
-            ChannelIdFixtures.CHANNEL_ID_4,
+            CHANNEL_ID_4,
             "BCash_Is_Trash",
             Pubkey.create("0298f6074a454a1f5345cb2a7c6f9fce206cd0bf675d177cdbf0ca7508dd28852f"),
             PoliciesDto.createFromModel(POLICIES_FOR_LOCAL_CHANNEL),
@@ -84,8 +88,8 @@ public class OpenChannelDtoFixture {
                     Coins.ofSatoshis(500)
             )));
 
-    public static final OpenChannelDto COTTO = new OpenChannelDto(
-            ChannelIdFixtures.CHANNEL_ID_5,
+    public static final OpenChannelDto C_OTTO = new OpenChannelDto(
+            CHANNEL_ID_5,
             "c-otto.de",
             Pubkey.create("027ce055380348d7812d2ae7745701c9f93e70c1adeb2657f053f91df4f2843c71"),
             PoliciesDto.createFromModel(POLICIES_FOR_LOCAL_CHANNEL),
